Add Austins_Terminal_Guide
Removed a significant amount of commands. Would prefer if it were more
verbose, but respecting authors work.

Also stripped down to keep it bare bones.

Added Austin and Paco to atuhor.json entries. Added pictures for each of
them, with Austin's being a Discord pic.

On the top of the Austins_Terminal_Guide.md it has a publish date and
also draft: true. Change this to false to have the page render for
testing, and if you want to have it appear under /posts in localhost
while testing make sure the publishDate has passed, though it will
render as long as draft: false.

Also started a small draft for advent_of_SCC.md for a final post on the 25th,
indexing the previous days of the year.