MATH + Post sc24-scc23-post-mortem comp days
Not well written. But just written to cover the days as I remember them.

Moreover it may be bland as it is just a timeline and not very
informative. Also am avoiding giving too much in takeaways because it
could be useful info I don't feel like sharing with outside groups.

Would like to have more pictures of the overall conference and other
stuff.

One can also do math now.
Either an equation as "$$ E = mc^2$$" or "\\(E = mc^2\\)" for inline
From: https://www.simonspavound.com/posts/2020/09/equations-with-katex-in-hugo/